refactor: ConfirmDialog 공통화 및 계정, 설정 모달 리팩터
