+++ frontend/src/components/commons/project-sidebar/AccountSettingsModalContent.tsx
@@ -10,29 +10,11 @@ import {
   TextFieldContent,
 } from '@wanteddev/wds';
 import { IconCircleCheckFill, IconClose, IconPencil } from '@wanteddev/wds-icon';
-import { ChangeEvent, useEffect, useRef, useState } from 'react';
+import type { ReactNode } from 'react';
 
-import { useDialog } from '@/components/commons/custom-dialog/DialogContext';
-
-import { ConfirmDialogContent } from './ConfirmDialogContent';
-
-// WDS TextField 포커스 테두리·캐럿 색을 FlowMeet Primary 토큰으로 스코프드 오버라이드
-const textFieldPrimaryFocusSx = {
-  '&:has(input:focus) [data-role="text-field-wrapper"]': {
-    boxShadow:
-      'inset 0 0 0 2px color-mix(in srgb, var(--color-primary-40) 43%, transparent) !important',
-  },
-  '[data-role="text-field-wrapper"] input': {
-    caretColor: 'var(--color-primary-40)',
-  },
-} as const;
-
-// WDS IconButton hover 시 배경 변화 제거
-const iconButtonNoHoverSx = {
-  '&:hover, &:focus-visible, &:active': {
-    backgroundColor: 'transparent !important',
-  },
-} as const;
+import { useAccountSettings } from '@/hooks/useAccountSettings';
+import { useProfileImagePreview } from '@/hooks/useProfileImagePreview';
+import { iconButtonNoHoverSx, textFieldPrimaryFocusSx } from '@/styles/sx';
 
 interface AccountSettingsModalContentProps {
   userName: string;
@@ -45,136 +27,51 @@ export const AccountSettingsModalContent = ({
   userEmail,
   onClose,
 }: AccountSettingsModalContentProps) => {
-  const { openDialog, closeDialog } = useDialog();
-  const [name, setName] = useState(userName);
-  const [email, setEmail] = useState(userEmail);
-  const [isEditingEmail, setIsEditingEmail] = useState(false);
-  const [verificationCode, setVerificationCode] = useState('');
-  const [isVerified, setIsVerified] = useState(false);
-  const [profilePreviewUrl, setProfilePreviewUrl] = useState<string | undefined>(undefined);
-  const objectUrlRef = useRef<string | null>(null);
+  const {
+    name,
+    setName,
+    email,
+    setEmail,
+    isEditingEmail,
+    verificationCode,
+    setVerificationCode,
+    isVerified,
+    handleStartEditEmail,
+    handleRequestVerification,
+    handleChangeEmail,
+    handleCloseClick,
+    handleLogoutClick,
+    handleWithdrawClick,
+  } = useAccountSettings({
+    initialName: userName,
+    initialEmail: userEmail,
+    onCloseModal: onClose,
+  });
+  const { previewUrl: profilePreviewUrl, handleImageChange: handleProfileImageChange } =
+    useProfileImagePreview();
 
-  useEffect(() => {
-    return () => {
-      if (objectUrlRef.current) {
-        URL.revokeObjectURL(objectUrlRef.current);
-      }
-    };
-  }, []);
-
-  const handleStartEditEmail = () => {
-    setIsEditingEmail(true);
-    setVerificationCode('');
-    setIsVerified(false);
-  };
-
-  // 데모 동작: 실제 인증 응답 대신 즉시 인증 성공으로 표시
-  const handleRequestVerification = () => {
-    setIsVerified(true);
-  };
-
-  const handleChangeEmail = () => {
-    setIsEditingEmail(false);
-  };
-
-  const handleCloseClick = () => {
-    if (!isEditingEmail) {
-      onClose();
-      return;
-    }
-    openDialog({
-      closeOnBackdrop: true,
-      closeOnEsc: true,
-      content: (
-        <ConfirmDialogContent
-          title="이메일 수정이 완료되지 않았어요"
-          description={'지금 나가면 입력한 내용이 저장되지 않을 수 있습니다.\n계속 등록하시겠어요?'}
-          leftButton={{
-            label: '취소하기',
-            tone: 'assistive',
-            onClick: () => {
-              closeDialog();
-              setIsEditingEmail(false);
-              setVerificationCode('');
-              setIsVerified(false);
-              onClose();
-            },
-          }}
-          rightButton={{
-            label: '계속하기',
-            tone: 'primary',
-            onClick: closeDialog,
-          }}
-        />
-      ),
-    });
-  };
-
-  const handleLogoutClick = () => {
-    openDialog({
-      closeOnBackdrop: true,
-      closeOnEsc: true,
-      content: (
-        <ConfirmDialogContent
-          title="로그아웃"
-          description="현재 계정에서 로그아웃할까요?"
-          leftButton={{
-            label: '로그아웃',
-            tone: 'assistive',
-            onClick: () => {
-              closeDialog();
-              onClose();
-            },
-          }}
-          rightButton={{
-            label: '취소',
-            tone: 'primary',
-            onClick: closeDialog,
-          }}
-        />
-      ),
-    });
-  };
-
-  const handleWithdrawClick = () => {
-    openDialog({
-      closeOnBackdrop: true,
-      closeOnEsc: true,
-      content: (
-        <ConfirmDialogContent
-          title="계정 탈퇴"
-          description={'정말로 계정을 탈퇴할까요?\n* 한 번 계정을 탈퇴하시면 복구가 불가능해요!'}
-          leftButton={{
-            label: '계정 삭제',
-            tone: 'negative',
-            onClick: () => {
-              closeDialog();
-              onClose();
-            },
-          }}
-          rightButton={{
-            label: '취소',
-            tone: 'assistive',
-            onClick: closeDialog,
-          }}
-        />
-      ),
-    });
-  };
-
-  const handleProfileImageChange = (event: ChangeEvent<HTMLInputElement>) => {
-    const file = event.target.files?.[0];
-    if (!file) return;
-
-    if (objectUrlRef.current) {
-      URL.revokeObjectURL(objectUrlRef.current);
-    }
-    const url = URL.createObjectURL(file);
-    objectUrlRef.current = url;
-    setProfilePreviewUrl(url);
-    // 동일 파일 재선택도 허용
-    event.target.value = '';
-  };
+  // 동일한 TextField에 편집 모드/읽기 모드 따라 trailing만 바꾼다.
+  // 읽기 모드에서는 readOnly 처리하여 "수정 버튼"을 눌러야만 편집이 가능하다.
+  const emailTrailing: ReactNode = isEditingEmail ? (
+    <TextFieldButton
+      variant="normal"
+      onClick={handleRequestVerification}
+      aria-label="이메일 인증 요청"
+    >
+      인증하기
+    </TextFieldButton>
+  ) : (
+    <TextFieldContent variant="icon-button">
+      <IconButton
+        variant="normal"
+        onClick={handleStartEditEmail}
+        aria-label="이메일 수정"
+        sx={iconButtonNoHoverSx}
+      >
+        <IconPencil />
+      </IconButton>
+    </TextFieldContent>
+  );
 
   return (
     <div className="flex w-full flex-col gap-8">
@@ -241,46 +138,20 @@ export const AccountSettingsModalContent = ({
               >
                 이메일
               </label>
-              {isEditingEmail ? (
-                <TextField
-                  id="account-settings-email"
-                  type="email"
-                  value={email}
-                  onChange={(event) => setEmail(event.target.value)}
-                  width="100%"
-                  sx={textFieldPrimaryFocusSx}
-                  trailingButton={
-                    <TextFieldButton
-                      variant="normal"
-                      onClick={handleRequestVerification}
-                      aria-label="이메일 인증 요청"
-                    >
-                      인증하기
-                    </TextFieldButton>
-                  }
-                />
-              ) : (
-                <TextField
-                  id="account-settings-email"
-                  type="email"
-                  value={email}
-                  onChange={(event) => setEmail(event.target.value)}
-                  width="100%"
-                  sx={textFieldPrimaryFocusSx}
-                  trailingContent={
-                    <TextFieldContent variant="icon-button">
-                      <IconButton
-                        variant="normal"
-                        onClick={handleStartEditEmail}
-                        aria-label="이메일 수정"
-                        sx={iconButtonNoHoverSx}
-                      >
-                        <IconPencil />
-                      </IconButton>
-                    </TextFieldContent>
-                  }
-                />
-              )}
+              {/* 편집 모드일 때만 입력이 가능하도록 readOnly로 잠근다.
+                  trailing 영역만 "수정 버튼" <-> "인증하기 버튼"으로 교체된다. */}
+              <TextField
+                id="account-settings-email"
+                type="email"
+                value={email}
+                onChange={(event) => setEmail(event.target.value)}
+                readOnly={!isEditingEmail}
+                width="100%"
+                sx={textFieldPrimaryFocusSx}
+                {...(isEditingEmail
+                  ? { trailingButton: emailTrailing }
+                  : { trailingContent: emailTrailing })}
+              />
             </div>
 
             {/* 편집 모드: 인증번호 입력 + 변경 버튼 */}
